fix: reservationId 오류 해결
