+++ Boolti/Boolti/Sources/UILayer/Concert/Ticketing/TicketingCompletion/TicketingCompletionViewController.swift
@@ -20,11 +20,13 @@ final class TicketingCompletionViewController: BooltiViewController {
     
     private let navigationBar = BooltiNavigationBar(type: .ticketingCompletion)
     
-    private var topContentView = UIView()
-    
-    private let depositSummaryView = DepositSummaryView()
-    
-    private let paymentCompletionView = PaymentCompletionView()
+    private let titleLabel: BooltiUILabel = {
+        let label = BooltiUILabel()
+        label.font = .aggroM(24)
+        label.textColor = .grey05
+        label.text = "결제를 완료했어요"
+        return label
+    }()
     
     private let underlineView: UIView = {
         let view = UIView()
@@ -33,15 +35,6 @@ final class TicketingCompletionViewController: BooltiViewController {
     }()
     
     private let reservedTicketView = ReservedTicketView()
-    
-    private let depositDetailView = DepositDetailView()
-    
-    private let copyButton: BooltiButton = {
-        let button = BooltiButton(title: "계좌번호 복사하기")
-        button.backgroundColor = .grey20
-        button.setTitleColor(.grey90, for: .normal)
-        return button
-    }()
     
     // MARK: Init
     
@@ -58,8 +51,6 @@ final class TicketingCompletionViewController: BooltiViewController {
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        
-        self.configureToastView(isButtonExisted: false)
         self.bindOutput()
         self.configureUI()
         self.configureConstraints()
@@ -96,40 +87,17 @@ extension TicketingCompletionViewController {
                 self.navigationController?.popToViewController(viewControllers[viewControllers.count - 3], animated: true)
             }
             .disposed(by: self.disposeBag)
-        
-        self.copyButton.rx.tap
-            .asDriver()
-            .drive(with: self) { owner, _ in
-                owner.viewModel.input.didCopyButtonTap.onNext(())
-                owner.showToast(message: "계좌번호가 복사되었어요")
-            }
-            .disposed(by: self.disposeBag)
     }
     
     private func bindOutput() {
         self.viewModel.ticketingData
             .take(1)
             .bind(with: self) { owner, data in
-                if data.selectedTicket.price == 0 {
-                    self.topContentView = self.paymentCompletionView
-                    [self.depositDetailView, self.copyButton].forEach { $0.isHidden = true }
-                } else {
-                    self.topContentView = self.depositSummaryView
-                }
-                
-                self.depositSummaryView.setData(date: data.concert.salesEndTime, price: data.selectedTicket.count * data.selectedTicket.price)
-                self.reservedTicketView.setData(concert: data.concert, selectedTicket: data.selectedTicket)
-                self.depositDetailView.setSalesEndTime(salesEndTime: data.concert.salesEndTime)
-            }
-            .disposed(by: self.disposeBag)
-        
-        self.viewModel.output.reservationDetail
-            .take(1)
-            .bind(with: self) { owner, data in
-                self.depositDetailView.setBankData(reservation: data)
+                owner.reservedTicketView.setData(concert: data.concert, selectedTicket: data.selectedTicket)
             }
             .disposed(by: self.disposeBag)
     }
+    
 }
 
 // MARK: - UI
@@ -138,11 +106,9 @@ extension TicketingCompletionViewController {
     
     private func configureUI() {
         self.view.addSubviews([self.navigationBar,
-                               self.topContentView,
+                               self.titleLabel,
                                self.underlineView,
-                               self.reservedTicketView,
-                               self.depositDetailView,
-                               self.copyButton])
+                               self.reservedTicketView])
         
         self.view.backgroundColor = .grey95
     }
@@ -154,13 +120,13 @@ extension TicketingCompletionViewController {
             make.horizontalEdges.equalToSuperview()
         }
         
-        self.topContentView.snp.makeConstraints { make in
-            make.top.equalTo(self.navigationBar.snp.bottom)
-            make.horizontalEdges.equalToSuperview()
+        self.titleLabel.snp.makeConstraints { make in
+            make.top.equalTo(self.navigationBar.snp.bottom).offset(20)
+            make.horizontalEdges.equalToSuperview().inset(20)
         }
         
         self.underlineView.snp.makeConstraints { make in
-            make.top.equalTo(self.topContentView.snp.bottom)
+            make.top.equalTo(self.titleLabel.snp.bottom).offset(20)
             make.horizontalEdges.equalToSuperview().inset(20)
             make.height.equalTo(1)
         }
@@ -169,15 +135,6 @@ extension TicketingCompletionViewController {
             make.top.equalTo(self.underlineView.snp.bottom)
             make.horizontalEdges.equalToSuperview()
         }
-        
-        self.depositDetailView.snp.makeConstraints { make in
-            make.top.equalTo(self.reservedTicketView.snp.bottom)
-            make.horizontalEdges.equalToSuperview().inset(20)
-        }
-        
-        self.copyButton.snp.makeConstraints { make in
-            make.top.equalTo(self.depositDetailView.snp.bottom).offset(16)
-            make.horizontalEdges.equalToSuperview().inset(20)
-        }
     }
+    
 }
